Add deep linking disable feature in settings (#190)
## Summary of the problem

The problem was that some users don't like the automatic opening of HCB
links in app.

## Describe your changes

This update introduces a new slider in the settings, allowing users to
choose whether links should open within the app or in their default
browser. When users opt out of HCB deep linking, the app now redirects
them to the original destination URL accordingly.

Due to the OS-level nature of iOS Universal Links, it's not possible to
completely disable deep linking via a simple setting. However, this
implementation includes a clever workaround: if deep linking is
disabled, the app briefly opens before automatically redirecting the
user to their intended destination. While this does result in a
momentary display of the app, the trade-off is worthwhile for the
improved user experience.

## Checklist

- [x] Descriptive PR title _(Does the title explain the changes in a
concise manner?)_
- [x] Tag related issues so they auto-close on merge
- [x] Easily digestible commits _(Are the commits small and easy to
understand?)_
[video](https://gist.github.com/garyhtou/97534180b0753aa607c35b6fdda9d2e0)
- [x] CI passes _(Do the GitHub checks pass?)_
- [x] Tested by submitter before requesting review _(Does it work in
development iOS/android? )_



https://github.com/user-attachments/assets/52c89825-8795-4f00-97f0-edec9bdede43