Add return grant button to grant cards (#175) (#189)
## Summary of the problem

Problem #175 - there is no button to return grants on the mobile app,
this adds one.

## Describe your changes

I've added a "Cancel Grant" button to each Grant Card. When clicked, it
displays a warning message—just like the one on HCB Web—informing the
user about the action they're about to take. The user can either cancel
or confirm the action. If confirmed, the grant is canceled using the new
HCB V4 API endpoint: `/api/v4/card_grants/:id/cancel`. A spinner appears
on the button to provide feedback to the client and prevent duplicate
API calls, and an error alert is shown if the API request fails.

## Checklist

- [x] Descriptive PR title _(Does the title explain the changes in a
concise manner?)_
- [x] Tag related issues so they auto-close on merge
- [x] Easily digestible commits _(Are the commits small and easy to
understand?)_
[video](https://gist.github.com/garyhtou/97534180b0753aa607c35b6fdda9d2e0)
- [x] CI passes _(Do the GitHub checks pass?)_
- [x] Tested by submitter before requesting review _(Does it work in
development iOS/android? )_



https://github.com/user-attachments/assets/ccb07d67-a6ac-4ac7-8893-019096009e28

---------

Co-authored-by: Mohamad Mortada <mmortada685@gmail.com>